backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
anthropic>=0.40.0
yfinance>=1.0.0
pycoingecko==3.1.0
python-dotenv==1.0.1
pydantic>=2.7.0
httpx>=0.27.0
websockets>=12.0
